add postgresql settings

+++ database.py
@@ -1,7 +1,16 @@
 import peewee
 from datetime import datetime, timedelta
+import configparser
 
-db = peewee.SqliteDatabase('todo.db')
+config = configparser.ConfigParser()
+config.read('./config.ini')
+
+db = peewee.PostgresqlDatabase(
+    config.get('postgres', 'database'),
+    user=config.get('postgres', 'user'),
+    password=config.get('postgres', 'password'),
+    host=config.get('postgres', 'host'),
+)
 
 
 class Todo(peewee.Model):
@@ -17,15 +26,17 @@ class Todo(peewee.Model):
 
 
 def create_table():
-    Todo.drop_table()
+    if Todo.table_exists():
+        None
+    else:
+        Todo.drop_table()
+        Todo.create_table()
 
-    Todo.create_table()
-
-    Todo.create(
-        id=1,
-        title='sample1',
-        details='details1',
-        done=False,
-        limit=datetime.now() + timedelta(days=1),
-        notification=datetime.now() + timedelta(days=0.8),
-    )
+        Todo.create(
+            id=1,
+            title='sample1',
+            details='details1',
+            done=False,
+            limit=datetime.now() + timedelta(days=1),
+            notification=datetime.now() + timedelta(days=0.8),
+        )
